Fix tablespace-based quota with the same role/schema
When there are two namespace_tablespace_quota configurations with the
same schema but different tablespaces, remove one quota will remove
another one automatically. The same bug happens for the
role_tabelspace_quota case. The root cause of this bug is that we
can't distinguish these two quotas using only quotaType and targetOid
of diskquota.quota_config.

This PR fix this by changing the meaning of targetOid column of
diskquota.quota_config table. When quotaType is
role_quota/schema_quota, targetOid refers to role_oid/schema_oid. When
quotaType is role_tablespace_quota/namespace_tablespace_quota,
targetOid refers to the rowId column of diskquota.target. This will
make it possible to distinguish two namespace_tablespace_quota with
the same schema and different tablespace. Because their rowId will be
different.

For the new quota type TABLESPACE_QUOTA. It is not a real quota, we
only use it for configuration. So we won't insert quotas with this
type into quota_info.